feat: implement RowActions component for transaction and receipt management
