+++ .vscode/settings.json
@@ -6,7 +6,6 @@
 	},
 	"oxc.typeAware": true,
 	"chat.tools.terminal.autoApprove": {
-		// Approve common, non-destructive commands
 		"/^ls/": true,
 		"/^cd/": true,
 		"/^grep/": true,
@@ -15,6 +14,7 @@
 		"npm run lint": true,
 		"npm run test": true,
 		"npm run check": true,
-		"npm run build": true
+		"npm run build": true,
+		"npx tsc": true
 	}
 }
