+++ src/routes/(app)/receipts/fuel/+page.server.ts
@@ -1,50 +1,22 @@
-import { error } from '@sveltejs/kit';
-import { superValidate } from 'sveltekit-superforms';
-import { zod4 } from 'sveltekit-superforms/adapters';
-
-import { transactionSchema } from '$lib/formSchemas';
-import { createCrudActions } from '$lib/server/actions/crud-helpers';
 import { transactionQueries } from '$lib/server/db/queries';
-import { transaction } from '$lib/server/db/schema';
-import { transactionBudgetAlertHooks } from '$lib/server/notifications/budget-threshold-alerts';
-import { formatDateForStorage, getMonthRangeFromUrl, getYearDateRange } from '$lib/utils/dates';
+import { getReceiptLoadContext, receiptActions } from '$lib/server/receipts/load-helpers';
 
 import type { PageServerLoad } from './$types';
 
 export const load: PageServerLoad = async ({ url, parent }) => {
-	// Get categories from parent layout
 	const { categories } = await parent();
-
-	// Find Gas category
-	const gasCategory = categories?.find((c) => c.name === 'Gas');
-	if (!gasCategory) {
-		throw error(404, 'Gas category not found');
-	}
-
-	// Get month, year, and date range from URL params
-	const { month, year, startDate, endDate } = getMonthRangeFromUrl(url);
-
-	// Get yearly date range
-	const { startDate: yearStartDate, endDate: yearEndDate } = getYearDateRange(year);
+	const { gasCategory, month, year, startDate, endDate, yearStartDate, yearEndDate, form } =
+		await getReceiptLoadContext(url, categories);
 
 	const currentDate = new Date();
 	const currentYear = currentDate.getFullYear();
 	const completedMonthsSinceJanuary =
 		year < currentYear ? 12 : year > currentYear ? 0 : currentDate.getMonth();
 
-	// Load monthly transactions for Gas category
-	const monthlyTransactions = await transactionQueries.findByCategory(gasCategory.id, {
-		start: startDate,
-		end: endDate
-	});
-
-	// Load yearly transactions for Gas category
-	const yearlyTransactions = await transactionQueries.findByCategory(gasCategory.id, {
-		start: yearStartDate,
-		end: yearEndDate
-	});
-
-	const form = await superValidate(zod4(transactionSchema));
+	const [monthlyTransactions, yearlyTransactions] = await Promise.all([
+		transactionQueries.findByCategory(gasCategory.id, { start: startDate, end: endDate }),
+		transactionQueries.findByCategory(gasCategory.id, { start: yearStartDate, end: yearEndDate })
+	]);
 
 	return {
 		monthlyTransactions,
@@ -56,26 +28,4 @@ export const load: PageServerLoad = async ({ url, parent }) => {
 	};
 };
 
-export const actions = createCrudActions({
-	schema: transactionSchema,
-	table: transaction,
-	entityName: 'Transaction',
-	...transactionBudgetAlertHooks,
-	transformCreate: (data, userId) => ({
-		amount: data.amount,
-		payee: data.payee,
-		notes: data.notes,
-		date: formatDateForStorage(data.date),
-		gstAmount: data.gstAmount ?? null,
-		categoryId: data.categoryId,
-		userId
-	}),
-	transformUpdate: (data) => ({
-		amount: data.amount,
-		payee: data.payee,
-		notes: data.notes,
-		date: formatDateForStorage(data.date),
-		gstAmount: data.gstAmount ?? null,
-		categoryId: data.categoryId
-	})
-});
+export const actions = receiptActions;
